Excel import treats an all missing column as numeric with NA, but we want it to be character. concerns columns missings and recode
keyImport with an NA in a string ov value_new will throwout the NA and replace it with "", which we probably don't want.

It should  be   key <- long2wide(wide2long(key))
but it is not workintg that way now, value_new "a|NA|b" comes back
"a||b".

I traced that behavior to end of long 2 wide, where I apparently
inserted this on purpose by a function printvals.   I don't know
what benefit I thought I was creating by that, but need to search
all usages of "long2wide" to try to find out.